refactor: remove unused login and deck components
- Deleted components: LoginForm, AIFlashcardsGenerator, ConfirmDeleteModal, CreateFlashcardModal, EditFlashcardModal, ReviewFlashcardModal, and FlashcardTable to streamline the codebase.
- Updated imports in DeckPage and HomePage to reflect new component structure.
- Improved modularity and maintainability by consolidating related components.
