+++ web/src/components/deckPage/FlashcardTable.jsx
