+++ web/src/components/loginPage/LoginForm.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../../contexts/AuthContext';
 import { Box, Button, TextField, Typography, Container, Alert } from '@mui/material';
 
 export const LoginForm = () => {
